+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Migrated `go.hacdias.com/indieauth` into `go.hacdias.com/indiekit/indieauth`.
+- Migrated `go.hacdias.com/indieauth` into `go.hacdias.com/indielib/indieauth`.
 - Added Micropub related tooling.
 
 ### Changed